+++ src/main/java/io/github/leanish/sqs/codec/CodecConfiguration.java
@@ -7,12 +7,14 @@ package io.github.leanish.sqs.codec;
 
 import io.github.leanish.sqs.codec.algorithms.ChecksumAlgorithm;
 import io.github.leanish.sqs.codec.algorithms.CompressionAlgorithm;
+import io.github.leanish.sqs.codec.algorithms.EncodingAlgorithm;
 
 /**
- * Immutable configuration for codec version, compression and checksum settings.
+ * Immutable configuration for codec version, compression, encoding and checksum settings.
  */
 public record CodecConfiguration(
         int version,
         CompressionAlgorithm compressionAlgorithm,
+        EncodingAlgorithm encodingAlgorithm,
         ChecksumAlgorithm checksumAlgorithm) {
 }